Rearrange bean instantiation and method calls
The bean instantiation of InjectToMe and method calls have been moved up to appear before the instantiation of InjectMe. This is because the InjectToMe bean needs to be created and its methods called first for processing order purposes. The same process was replicated for the second instantiation of the InjectToMe bean.