Shift setting of custom message in dependency injection demo
Refactored the code to move the setting of custom message from injectMe to injectMe2 in SpringBootDependencyInjectionDemoApplication. This is in line with demonstrating the singleton behaviour of beans in Spring, i.e., changing the property of one should reflect in others. The change also helps in ensuring consistency of displayed messages across all InjectMe beans.